+++ sdk/tables/data-tables/recordings/node/batch_operations/recording_should_handle_sub_request_error.js
@@ -1,13 +1,13 @@
 let nock = require('nock');
 
-module.exports.hash = "efabed6507d7b7683d890e6343a3a3bc";
+module.exports.hash = "095b0df362273370b396798642a53e62";
 
 module.exports.testInfo = {"uniqueName":{},"newDate":{}}
 
 nock('https://fakestorageaccount.table.core.windows.net:443', {"encodedQueryParams":true})
   .post('/Tables', {"TableName":"batchTableTestnode"})
   .query(true)
-  .reply(409, {"odata.error":{"code":"TableAlreadyExists","message":{"lang":"en-US","value":"The table specified already exists.\nRequestId:a1f0788a-b002-0072-5276-98d8c7000000\nTime:2020-10-02T04:44:43.9542700Z"}}}, [
+  .reply(409, {"odata.error":{"code":"TableAlreadyExists","message":{"lang":"en-US","value":"The table specified already exists.\nRequestId:8a1d10c0-0002-006b-3511-99f4af000000\nTime:2020-10-02T23:12:23.0807201Z"}}}, [
   'Cache-Control',
   'no-cache',
   'Transfer-Encoding',
@@ -17,37 +17,37 @@ nock('https://fakestorageaccount.table.core.windows.net:443', {"encodedQueryPara
   'Server',
   'Windows-Azure-Table/1.0 Microsoft-HTTPAPI/2.0',
   'x-ms-request-id',
-  'a1f0788a-b002-0072-5276-98d8c7000000',
+  '8a1d10c0-0002-006b-3511-99f4af000000',
   'x-ms-client-request-id',
-  '067b5d73-ea5f-4960-8a25-d91c31a8bab6',
+  '8a992ff2-61e9-45d2-89a3-724706fccadb',
   'x-ms-version',
   '2019-02-02',
   'X-Content-Type-Options',
   'nosniff',
   'Date',
-  'Fri, 02 Oct 2020 04:44:43 GMT'
+  'Fri, 02 Oct 2020 23:12:22 GMT'
 ]);
 
 nock('https://fakestorageaccount.table.core.windows.net:443', {"encodedQueryParams":true})
-  .post('/$batch', "--batch_mockBatchId\r\ncontent-type: multipart/mixed; boundary=changeset_mockChangesetId\r\n\r\n\r\n--changeset_mockChangesetId\r\ncontent-type: application/http\r\ncontent-transfer-encoding: binary\r\n\r\nPOST https://fakestorageaccount.table.core.windows.net/noExistingTable HTTP/1.1\r\nContent-Type: application/json;odata=nometadata\r\nDataServiceVersion: 3.0\r\nAccept: application/json;odata=minimalmetadata\r\nPrefer: return-no-content\r\n\r\n{\"PartitionKey\":\"batchTest\",\"RowKey\":\"1\",\"name\":\"first\"}\r\n--changeset_mockChangesetId\r\ncontent-type: application/http\r\ncontent-transfer-encoding: binary\r\n\r\nPOST https://fakestorageaccount.table.core.windows.net/noExistingTable HTTP/1.1\r\nContent-Type: application/json;odata=nometadata\r\nDataServiceVersion: 3.0\r\nAccept: application/json;odata=minimalmetadata\r\nPrefer: return-no-content\r\n\r\n{\"PartitionKey\":\"batchTest\",\"RowKey\":\"2\",\"name\":\"second\"}\r\n--changeset_mockChangesetId\r\ncontent-type: application/http\r\ncontent-transfer-encoding: binary\r\n\r\nPOST https://fakestorageaccount.table.core.windows.net/noExistingTable HTTP/1.1\r\nContent-Type: application/json;odata=nometadata\r\nDataServiceVersion: 3.0\r\nAccept: application/json;odata=minimalmetadata\r\nPrefer: return-no-content\r\n\r\n{\"PartitionKey\":\"batchTest\",\"RowKey\":\"3\",\"name\":\"third\"}\r\n--changeset_mockChangesetId--\r\n--batch_mockBatchId\r\n")
+  .post('/$batch', "--batch_fakeId\r\ncontent-type: multipart/mixed; boundary=changeset_fakeId\r\n\r\n\r\n--changeset_fakeId\r\ncontent-type: application/http\r\ncontent-transfer-encoding: binary\r\n\r\nPOST https://fakestorageaccount.table.core.windows.net/noExistingTable HTTP/1.1\r\nContent-Type: application/json;odata=nometadata\r\nDataServiceVersion: 3.0\r\nAccept: application/json;odata=minimalmetadata\r\nPrefer: return-no-content\r\n\r\n{\"PartitionKey\":\"batchTest\",\"RowKey\":\"1\",\"name\":\"first\"}\r\n--changeset_fakeId\r\ncontent-type: application/http\r\ncontent-transfer-encoding: binary\r\n\r\nPOST https://fakestorageaccount.table.core.windows.net/noExistingTable HTTP/1.1\r\nContent-Type: application/json;odata=nometadata\r\nDataServiceVersion: 3.0\r\nAccept: application/json;odata=minimalmetadata\r\nPrefer: return-no-content\r\n\r\n{\"PartitionKey\":\"batchTest\",\"RowKey\":\"2\",\"name\":\"second\"}\r\n--changeset_fakeId\r\ncontent-type: application/http\r\ncontent-transfer-encoding: binary\r\n\r\nPOST https://fakestorageaccount.table.core.windows.net/noExistingTable HTTP/1.1\r\nContent-Type: application/json;odata=nometadata\r\nDataServiceVersion: 3.0\r\nAccept: application/json;odata=minimalmetadata\r\nPrefer: return-no-content\r\n\r\n{\"PartitionKey\":\"batchTest\",\"RowKey\":\"3\",\"name\":\"third\"}\r\n--changeset_fakeId--\r\n--batch_fakeId\r\n")
   .query(true)
-  .reply(202, "--batchresponse_78d1f87d-f77f-4c3d-ba0e-e84f1627fd44\r\nContent-Type: multipart/mixed; boundary=changesetresponse_77edc21f-2b59-4d25-8dc0-7427c7a25e40\r\n\r\n--changesetresponse_77edc21f-2b59-4d25-8dc0-7427c7a25e40\r\nContent-Type: application/http\r\nContent-Transfer-Encoding: binary\r\n\r\nHTTP/1.1 404 Not Found\r\nX-Content-Type-Options: nosniff\r\nDataServiceVersion: 3.0;\r\nContent-Type: application/json;odata=minimalmetadata;streaming=true;charset=utf-8\r\n\r\n{\"odata.error\":{\"code\":\"TableNotFound\",\"message\":{\"lang\":\"en-US\",\"value\":\"0:The table specified does not exist.\\nRequestId:678247da-6002-003f-5376-981e25000000\\nTime:2020-10-02T04:44:44.0585931Z\"}}}\r\n--changesetresponse_77edc21f-2b59-4d25-8dc0-7427c7a25e40--\r\n--batchresponse_78d1f87d-f77f-4c3d-ba0e-e84f1627fd44--\r\n", [
+  .reply(202, "--batchresponse_b836c2da-e5cd-4392-a662-41fac40bc137\r\nContent-Type: multipart/mixed; boundary=changesetresponse_93e24565-e456-4bd1-917f-246c970a4614\r\n\r\n--changesetresponse_93e24565-e456-4bd1-917f-246c970a4614\r\nContent-Type: application/http\r\nContent-Transfer-Encoding: binary\r\n\r\nHTTP/1.1 404 Not Found\r\nX-Content-Type-Options: nosniff\r\nDataServiceVersion: 3.0;\r\nContent-Type: application/json;odata=minimalmetadata;streaming=true;charset=utf-8\r\n\r\n{\"odata.error\":{\"code\":\"TableNotFound\",\"message\":{\"lang\":\"en-US\",\"value\":\"0:The table specified does not exist.\\nRequestId:9111bcc8-a002-0044-0211-997595000000\\nTime:2020-10-02T23:12:23.1938601Z\"}}}\r\n--changesetresponse_93e24565-e456-4bd1-917f-246c970a4614--\r\n--batchresponse_b836c2da-e5cd-4392-a662-41fac40bc137--\r\n", [
   'Cache-Control',
   'no-cache',
   'Transfer-Encoding',
   'chunked',
   'Content-Type',
-  'multipart/mixed; boundary=batchresponse_78d1f87d-f77f-4c3d-ba0e-e84f1627fd44',
+  'multipart/mixed; boundary=batchresponse_b836c2da-e5cd-4392-a662-41fac40bc137',
   'Server',
   'Windows-Azure-Table/1.0 Microsoft-HTTPAPI/2.0',
   'x-ms-request-id',
-  '678247da-6002-003f-5376-981e25000000',
+  '9111bcc8-a002-0044-0211-997595000000',
   'x-ms-client-request-id',
-  'a18f18a1-1c74-44c5-b3eb-eb4790cdacd9',
+  '9f34f8db-2e32-4197-b565-0d6dc476b61e',
   'x-ms-version',
   '2019-02-02',
   'X-Content-Type-Options',
   'nosniff',
   'Date',
-  'Fri, 02 Oct 2020 04:44:43 GMT'
+  'Fri, 02 Oct 2020 23:12:22 GMT'
 ]);