+++ sdk/tables/data-tables/recordings/node/batch_operations/recording_should_send_a_set_of_delete_batch_operations.js
@@ -1,13 +1,13 @@
 let nock = require('nock');
 
-module.exports.hash = "19faa959b3c40d8f8b5f217ad855c37e";
+module.exports.hash = "d10683902da795c01916add792ff5e87";
 
 module.exports.testInfo = {"uniqueName":{},"newDate":{}}
 
 nock('https://fakestorageaccount.table.core.windows.net:443', {"encodedQueryParams":true})
   .post('/Tables', {"TableName":"batchTableTestnode"})
   .query(true)
-  .reply(409, {"odata.error":{"code":"TableAlreadyExists","message":{"lang":"en-US","value":"The table specified already exists.\nRequestId:a29189b4-8002-0053-7f76-98b5f6000000\nTime:2020-10-02T04:44:43.7153166Z"}}}, [
+  .reply(409, {"odata.error":{"code":"TableAlreadyExists","message":{"lang":"en-US","value":"The table specified already exists.\nRequestId:933285f3-b002-0036-8011-9904ab000000\nTime:2020-10-02T23:12:22.8490292Z"}}}, [
   'Cache-Control',
   'no-cache',
   'Transfer-Encoding',
@@ -17,37 +17,37 @@ nock('https://fakestorageaccount.table.core.windows.net:443', {"encodedQueryPara
   'Server',
   'Windows-Azure-Table/1.0 Microsoft-HTTPAPI/2.0',
   'x-ms-request-id',
-  'a29189b4-8002-0053-7f76-98b5f6000000',
+  '933285f3-b002-0036-8011-9904ab000000',
   'x-ms-client-request-id',
-  'a45e3e39-e3cb-4a7d-bd0c-77f2fa48cf71',
+  'c42a6101-9a7e-46f3-a883-e4a130ce6768',
   'x-ms-version',
   '2019-02-02',
   'X-Content-Type-Options',
   'nosniff',
   'Date',
-  'Fri, 02 Oct 2020 04:44:42 GMT'
+  'Fri, 02 Oct 2020 23:12:22 GMT'
 ]);
 
 nock('https://fakestorageaccount.table.core.windows.net:443', {"encodedQueryParams":true})
-  .post('/$batch', "--batch_mockBatchId\r\ncontent-type: multipart/mixed; boundary=changeset_mockChangesetId\r\n\r\n\r\n--changeset_mockChangesetId\r\ncontent-type: application/http\r\ncontent-transfer-encoding: binary\r\n\r\nDELETE https://fakestorageaccount.table.core.windows.net/batchTableTestnode(PartitionKey='batchTest',RowKey='1') HTTP/1.1\r\nContent-Type: application/json; charset=utf-8\r\nAccept: application/json;odata=minimalmetadata\r\nDataServiceVersion: 3.0\r\nIf-Match: *\r\n\r\nundefined\r\n--changeset_mockChangesetId\r\ncontent-type: application/http\r\ncontent-transfer-encoding: binary\r\n\r\nDELETE https://fakestorageaccount.table.core.windows.net/batchTableTestnode(PartitionKey='batchTest',RowKey='2') HTTP/1.1\r\nContent-Type: application/json; charset=utf-8\r\nAccept: application/json;odata=minimalmetadata\r\nDataServiceVersion: 3.0\r\nIf-Match: *\r\n\r\nundefined\r\n--changeset_mockChangesetId\r\ncontent-type: application/http\r\ncontent-transfer-encoding: binary\r\n\r\nDELETE https://fakestorageaccount.table.core.windows.net/batchTableTestnode(PartitionKey='batchTest',RowKey='3') HTTP/1.1\r\nContent-Type: application/json; charset=utf-8\r\nAccept: application/json;odata=minimalmetadata\r\nDataServiceVersion: 3.0\r\nIf-Match: *\r\n\r\nundefined\r\n--changeset_mockChangesetId--\r\n--batch_mockBatchId\r\n")
+  .post('/$batch', "--batch_fakeId\r\ncontent-type: multipart/mixed; boundary=changeset_fakeId\r\n\r\n\r\n--changeset_fakeId\r\ncontent-type: application/http\r\ncontent-transfer-encoding: binary\r\n\r\nDELETE https://fakestorageaccount.table.core.windows.net/batchTableTestnode(PartitionKey='batchTest',RowKey='1') HTTP/1.1\r\nContent-Type: application/json; charset=utf-8\r\nAccept: application/json;odata=minimalmetadata\r\nDataServiceVersion: 3.0\r\nIf-Match: *\r\n\r\nundefined\r\n--changeset_fakeId\r\ncontent-type: application/http\r\ncontent-transfer-encoding: binary\r\n\r\nDELETE https://fakestorageaccount.table.core.windows.net/batchTableTestnode(PartitionKey='batchTest',RowKey='2') HTTP/1.1\r\nContent-Type: application/json; charset=utf-8\r\nAccept: application/json;odata=minimalmetadata\r\nDataServiceVersion: 3.0\r\nIf-Match: *\r\n\r\nundefined\r\n--changeset_fakeId\r\ncontent-type: application/http\r\ncontent-transfer-encoding: binary\r\n\r\nDELETE https://fakestorageaccount.table.core.windows.net/batchTableTestnode(PartitionKey='batchTest',RowKey='3') HTTP/1.1\r\nContent-Type: application/json; charset=utf-8\r\nAccept: application/json;odata=minimalmetadata\r\nDataServiceVersion: 3.0\r\nIf-Match: *\r\n\r\nundefined\r\n--changeset_fakeId--\r\n--batch_fakeId\r\n")
   .query(true)
-  .reply(202, "--batchresponse_5b7783f2-9631-44d9-90d4-582740f22696\r\nContent-Type: multipart/mixed; boundary=changesetresponse_883ee6fc-afa7-41ea-a4b1-dd70c8ec8081\r\n\r\n--changesetresponse_883ee6fc-afa7-41ea-a4b1-dd70c8ec8081\r\nContent-Type: application/http\r\nContent-Transfer-Encoding: binary\r\n\r\nHTTP/1.1 204 No Content\r\nX-Content-Type-Options: nosniff\r\nCache-Control: no-cache\r\nDataServiceVersion: 1.0;\r\n\r\n\r\n--changesetresponse_883ee6fc-afa7-41ea-a4b1-dd70c8ec8081\r\nContent-Type: application/http\r\nContent-Transfer-Encoding: binary\r\n\r\nHTTP/1.1 204 No Content\r\nX-Content-Type-Options: nosniff\r\nCache-Control: no-cache\r\nDataServiceVersion: 1.0;\r\n\r\n\r\n--changesetresponse_883ee6fc-afa7-41ea-a4b1-dd70c8ec8081\r\nContent-Type: application/http\r\nContent-Transfer-Encoding: binary\r\n\r\nHTTP/1.1 204 No Content\r\nX-Content-Type-Options: nosniff\r\nCache-Control: no-cache\r\nDataServiceVersion: 1.0;\r\n\r\n\r\n--changesetresponse_883ee6fc-afa7-41ea-a4b1-dd70c8ec8081--\r\n--batchresponse_5b7783f2-9631-44d9-90d4-582740f22696--\r\n", [
+  .reply(202, "--batchresponse_16735c58-84cc-4d94-a1d1-42b13560c0ee\r\nContent-Type: multipart/mixed; boundary=changesetresponse_f4d4d340-7b12-4be9-85e8-e8aa5797b426\r\n\r\n--changesetresponse_f4d4d340-7b12-4be9-85e8-e8aa5797b426\r\nContent-Type: application/http\r\nContent-Transfer-Encoding: binary\r\n\r\nHTTP/1.1 204 No Content\r\nX-Content-Type-Options: nosniff\r\nCache-Control: no-cache\r\nDataServiceVersion: 1.0;\r\n\r\n\r\n--changesetresponse_f4d4d340-7b12-4be9-85e8-e8aa5797b426\r\nContent-Type: application/http\r\nContent-Transfer-Encoding: binary\r\n\r\nHTTP/1.1 204 No Content\r\nX-Content-Type-Options: nosniff\r\nCache-Control: no-cache\r\nDataServiceVersion: 1.0;\r\n\r\n\r\n--changesetresponse_f4d4d340-7b12-4be9-85e8-e8aa5797b426\r\nContent-Type: application/http\r\nContent-Transfer-Encoding: binary\r\n\r\nHTTP/1.1 204 No Content\r\nX-Content-Type-Options: nosniff\r\nCache-Control: no-cache\r\nDataServiceVersion: 1.0;\r\n\r\n\r\n--changesetresponse_f4d4d340-7b12-4be9-85e8-e8aa5797b426--\r\n--batchresponse_16735c58-84cc-4d94-a1d1-42b13560c0ee--\r\n", [
   'Cache-Control',
   'no-cache',
   'Transfer-Encoding',
   'chunked',
   'Content-Type',
-  'multipart/mixed; boundary=batchresponse_5b7783f2-9631-44d9-90d4-582740f22696',
+  'multipart/mixed; boundary=batchresponse_16735c58-84cc-4d94-a1d1-42b13560c0ee',
   'Server',
   'Windows-Azure-Table/1.0 Microsoft-HTTPAPI/2.0',
   'x-ms-request-id',
-  'e7212b3e-2002-0033-3e76-98f0d4000000',
+  'fde6b0d1-f002-0018-6411-99846c000000',
   'x-ms-client-request-id',
-  'aee57a68-341c-4d98-b064-3e124e71220d',
+  '553cdfcb-7af6-463a-9781-999ae3dd865c',
   'x-ms-version',
   '2019-02-02',
   'X-Content-Type-Options',
   'nosniff',
   'Date',
-  'Fri, 02 Oct 2020 04:44:43 GMT'
+  'Fri, 02 Oct 2020 23:12:22 GMT'
 ]);