+++ sdk/tables/data-tables/src/models.ts
@@ -12,14 +12,6 @@ import {
   TableCreateHeaders
 } from "./generated/models";
 import { OperationOptions, HttpResponse, PipelineOptions } from "@azure/core-http";
-import {
-  DeleteTableOptions,
-  DeleteTableResponse,
-  DeleteTableEntityResponse,
-  UpdateEntityResponse,
-  UpsertEntityResponse
-} from "./generatedModels";
-import { PagedAsyncIterableIterator } from "@azure/core-paging";
 
 /**
  * Client options used to configure Tables Api requests
@@ -446,97 +438,6 @@ export interface TableBatch {
 }
 
 /**
- * Describes the shape of a TableClient
- */
-export interface TableClientLike {
-  /**
-   * Name of the table to perform operations on.
-   */
-  readonly tableName: string;
-  /**
-   *  Creates the current table it it doesn't exist
-   * @param options The options parameters.
-   */
-  create(options?: CreateTableOptions): Promise<CreateTableItemResponse>;
-  /**
-   * Creates a new Batch to collect sub-operations that can be submitted together via submitBatch
-   * @param partitionKey partitionKey to which the batch operations will be targetted to
-   */
-  createBatch(partitionKey: string): TableBatch;
-  /**
-   * Insert entity in the table.
-   * @param entity The properties for the table entity.
-   * @param options The options parameters.
-   */
-  createEntity<T extends object>(
-    entity: TableEntity<T>,
-    options?: CreateTableEntityOptions
-  ): Promise<CreateTableEntityResponse>;
-  /**
-   * Permanently deletes the current table with all of its entities.
-   * @param options The options parameters.
-   */
-  delete(options?: DeleteTableOptions): Promise<DeleteTableResponse>;
-  /**
-   * Deletes the specified entity in the table.
-   * @param partitionKey The partition key of the entity.
-   * @param rowKey The row key of the entity.
-   * @param options The options parameters.
-   */
-  deleteEntity(
-    partitionKey: string,
-    rowKey: string,
-    options?: DeleteTableEntityOptions
-  ): Promise<DeleteTableEntityResponse>;
-  /**
-   * Returns a single entity in the table.
-   * @param partitionKey The partition key of the entity.
-   * @param rowKey The row key of the entity.
-   * @param options The options parameters.
-   */
-  getEntity<T extends object>(
-    partitionKey: string,
-    rowKey: string,
-    options?: GetTableEntityOptions
-  ): Promise<GetTableEntityResponse<T>>;
-  /**
-   * Queries entities in a table.
-   * @param tableName The name of the table.
-   * @param options The options parameters.
-   */
-  listEntities<T extends object>(
-    options?: ListTableEntitiesOptions
-  ): PagedAsyncIterableIterator<T, ListEntitiesResponse<T>>;
-  /**
-   * Update an entity in the table.
-   * @param entity The properties of the entity to be updated.
-   * @param mode The different modes for updating the entity:
-   *             - Merge: Updates an entity by updating the entity's properties without replacing the existing entity.
-   *             - Replace: Updates an existing entity by replacing the entire entity.
-   * @param options The options parameters.
-   */
-  updateEntity<T extends object>(
-    entity: TableEntity<T>,
-    mode: UpdateMode,
-    options?: UpdateTableEntityOptions
-  ): Promise<UpdateEntityResponse>;
-  /**
-   * Upsert an entity in the table.
-   * @param tableName The name of the table.
-   * @param entity The properties for the table entity.
-   * @param mode The different modes for updating the entity:
-   *             - Merge: Updates an entity by updating the entity's properties without replacing the existing entity.
-   *             - Replace: Updates an existing entity by replacing the entire entity.
-   * @param options The options parameters.
-   */
-  upsertEntity<T extends object>(
-    entity: TableEntity<T>,
-    mode: UpdateMode,
-    options?: UpsertTableEntityOptions
-  ): Promise<UpsertEntityResponse>;
-}
-
-/**
  * Represents the response of a Batch operation
  */
 export interface TableBatchResponse {
@@ -570,22 +471,4 @@ export interface TableBatchEntityResponse {
    * Sub-response status
    */
   status: number;
-  /**
-   * Body content of the sub-request
-   */
-  body?: any;
-}
-
-/**
- * Options for TableBatch creation
- */
-export interface TableBatchOptions {
-  /**
-   * Changeset Id. If not provided a new UUID will be assigned internally
-   */
-  changesetId?: string;
-  /**
-   * Batch Id. If not provided a new UUID will be assigned internally
-   */
-  batchId?: string;
 }